Improved documentation and minor refactoring
